requirements.txt
aiofiles==0.6.0
kademlia==2.2.2
websockets==9.0.1
py-multihash
py-cid==0.3.0
py-multicodec==0.2.1
